requirements.txt
torch>=2.0
torchvision
timm
scipy
numpy
tqdm
pyyaml
optuna
Pillow
tensorflow>=2.10
tensorflow-datasets